# contracts/api/platform-control.openapi.yaml
...
openapi: 3.1.0
info:
  title: Platform Control API
  description: |
    API for managing sources, source versions, runs, approvals, and provider webhooks
    in the Evidara platform control plane.

    This document is generated from the FastAPI application — it is exactly what the
    service serves at `/openapi.json`. Do not hand-edit `contracts/api/platform-control.openapi.yaml`;
    change the models or routes and run `scripts/generate_platform_control_contract.py`.
  contact:
    name: Evidara Team
  version: 0.14.0
paths:
...
  /v1/runs:
    get:
      tags:
      - runs
      - agent-discovery
      summary: List Runs
      operationId: listRuns
      security:
      - APIKeyHeader: []
      parameters:
      - name: mode
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/RunMode'
          - type: 'null'
          title: Mode
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/RunStatus'
          - type: 'null'
          title: Status
      - name: source_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Id
      - name: refused
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          description: Filter on ADR-0030 two-key-lock refusals (#634). `true` returns only refused dispatches
            — what an operator tried to onboard and why it was blocked; `false` excludes them so they
            do not pollute failure triage.
          title: Refused
        description: Filter on ADR-0030 two-key-lock refusals (#634). `true` returns only refused dispatches
          — what an operator tried to onboard and why it was blocked; `false` excludes them so they do
          not pollute failure triage.
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 100
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - runs
      summary: Create Run
      operationId: createRun
      security:
      - APIKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRunRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunResponse'
        '400':
          description: A domain rule rejected the request (misconfigured provider, orchestration failure,
            or a blueprint template that is not enabled for live acquisition).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: A referenced entity does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The request conflicts with the current state of the resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/runs/readiness:
    get:
      tags:
      - runs
      summary: Get Run Readiness
      operationId: getRunReadiness
      security:
      - APIKeyHeader: []
      parameters:
      - name: source_id
        in: query
        required: true
        schema:
          type: string
          title: Source Id
      - name: source_version_id
        in: query
        required: true
        schema:
          type: string
          title: Source Version Id
      - name: mode
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/RunMode'
          default: preview
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunReadinessResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/runs/{run_id}:
    get:
      tags:
      - runs
      - agent-discovery
      summary: Get Run
      operationId: getRun
      security:
      - APIKeyHeader: []
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunResponse'
        '404':
          description: A referenced entity does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/runs/{run_id}/captured-resources:
    get:
      tags:
      - runs
      - agent-discovery
      summary: List Run Captured Resources
      operationId: listRunCapturedResources
      security:
      - APIKeyHeader: []
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapturedResourceListResponse'
        '404':
          description: A referenced entity does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/runs/{run_id}/raw-artifacts:
    get:
      tags:
      - runs
      - agent-discovery
      summary: List Run Raw Artifacts
      operationId: listRunRawArtifacts
      security:
      - APIKeyHeader: []
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RawArtifactListResponse'
        '404':
          description: A referenced entity does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/runs/{run_id}/provider-jobs:
    get:
      tags:
      - runs
      - agent-discovery
      summary: List Run Provider Jobs
      operationId: listRunProviderJobs
      security:
      - APIKeyHeader: []
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderJobListResponse'
        '404':
          description: A referenced entity does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/runs/{run_id}/cancel:
    post:
      tags:
      - runs
      summary: Cancel Run
      operationId: cancelRun
      security:
      - APIKeyHeader: []
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunResponse'
        '404':
          description: A referenced entity does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The request conflicts with the current state of the resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/runs/{run_id}/retry:
    post:
      tags:
      - runs
      summary: Retry Run
      description: Retry a failed or cancelled run. Resets the run to PENDING status.
      operationId: retryRun
      security:
      - APIKeyHeader: []
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunResponse'
        '400':
          description: A domain rule rejected the request (misconfigured provider, orchestration failure,
            or a blueprint template that is not enabled for live acquisition).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: A referenced entity does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The request conflicts with the current state of the resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/runs/{run_id}/preview-summary:
    get:
      tags:
      - runs
      - agent-discovery
      summary: Get Run Preview Summary
      operationId: getRunPreviewSummary
      security:
      - APIKeyHeader: []
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunPreviewSummaryResponse'
        '404':
          description: A referenced entity does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/runs/{run_id}/processing-status:
    get:
      tags:
      - runs
      - agent-discovery
      summary: List Run Processing Status
      operationId: listRunProcessingStatus
      security:
      - APIKeyHeader: []
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingStatusUpdateListResponse'
        '404':
          description: A referenced entity does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/runs/{run_id}/document-lifecycle:
    get:
      tags:
      - runs
      - agent-discovery
      summary: List Run Document Lifecycle
      operationId: listRunDocumentLifecycle
      security:
      - APIKeyHeader: []
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentLifecycleEventListResponse'
        '404':
          description: A referenced entity does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/runs/{run_id}/lifecycle:
    get:
      tags:
      - runs
      summary: Get Run Lifecycle
      description: 'Aggregated timeline: run events, processing status, and document lifecycle.'
      operationId: getRunLifecycle
      security:
      - APIKeyHeader: []
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunLifecycleResponse'
        '404':
          description: A referenced entity does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/wizard/runs/{run_id}:
    get:
      tags:
      - wizard
      summary: Get Wizard Run
      operationId: getWizardRun
      security:
      - APIKeyHeader: []
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WizardRunStatusResponse'
        '404':
          description: A referenced entity does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CapturedResourceListResponse:
      properties:
        data:
          items:
            $ref: '#/components/schemas/CapturedResourceResponse'
          type: array
          title: Data
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - data
      - total
      - limit
      - offset
      title: CapturedResourceListResponse
...
    CreateRunRequest:
      properties:
        source_id:
          type: string
          title: Source Id
        source_version_id:
          type: string
          title: Source Version Id
        mode:
          $ref: '#/components/schemas/RunMode'
          default: preview
        scope:
          $ref: '#/components/schemas/RunScopeRequest'
        replay:
          anyOf:
          - $ref: '#/components/schemas/RunReplayRequest'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - source_id
      - source_version_id
      title: CreateRunRequest
...
    DocumentLifecycleEventListResponse:
      properties:
        data:
          items:
            $ref: '#/components/schemas/DocumentLifecycleEventResponse'
          type: array
          title: Data
      type: object
      required:
      - data
      title: DocumentLifecycleEventListResponse
...
    ErrorResponse:
      properties:
        detail:
          type: string
          title: Detail
          description: Human-readable explanation of the failure.
        correlation_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Correlation Id
          description: Correlation id for the request, echoed as the X-Correlation-Id header.
      type: object
      required:
      - detail
      title: ErrorResponse
      description: Uniform error body. `correlation_id` echoes `X-Correlation-Id` when present.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ProcessingStatusUpdateListResponse:
      properties:
        data:
          items:
            $ref: '#/components/schemas/ProcessingStatusUpdateResponse'
          type: array
          title: Data
      type: object
      required:
      - data
      title: ProcessingStatusUpdateListResponse
...
    ProviderJobListResponse:
      properties:
        data:
          items:
            $ref: '#/components/schemas/ProviderJobResponse'
          type: array
          title: Data
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - data
      - total
      - limit
      - offset
      title: ProviderJobListResponse
...
    RawArtifactListResponse:
      properties:
        data:
          items:
            $ref: '#/components/schemas/RawArtifactResponse'
          type: array
          title: Data
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - data
      - total
      - limit
      - offset
      title: RawArtifactListResponse
...
    RunListResponse:
      properties:
        data:
          items:
            $ref: '#/components/schemas/RunListItemResponse'
          type: array
          title: Data
        total:
          anyOf:
          - type: integer
          - type: 'null'
          title: Total
        limit:
          anyOf:
          - type: integer
          - type: 'null'
          title: Limit
        offset:
          anyOf:
          - type: integer
          - type: 'null'
          title: Offset
      type: object
      required:
      - data
      title: RunListResponse
    RunMode:
      type: string
      enum:
      - preview
      - production
      title: RunMode
...
    RunPreviewSummaryResponse:
      properties:
        run_id:
          type: string
          title: Run Id
        captured_url_count:
          type: integer
          title: Captured Url Count
        artifacts_count:
          type: integer
          title: Artifacts Count
        captured_resources_count:
          type: integer
          title: Captured Resources Count
        pdf_count:
          type: integer
          title: Pdf Count
        likely_decision_page_count:
          type: integer
          title: Likely Decision Page Count
        likely_boilerplate_page_count:
          type: integer
          title: Likely Boilerplate Page Count
        likely_duplicate_page_count:
          type: integer
          title: Likely Duplicate Page Count
        content_type_breakdown:
          items:
            $ref: '#/components/schemas/RunPreviewSummaryBreakdownEntry'
          type: array
          title: Content Type Breakdown
        likely_decision_pages:
          items:
            $ref: '#/components/schemas/RunPreviewSummarySample'
          type: array
          title: Likely Decision Pages
        likely_boilerplate_pages:
          items:
            $ref: '#/components/schemas/RunPreviewSummarySample'
          type: array
          title: Likely Boilerplate Pages
        likely_duplicate_pages:
          items:
            $ref: '#/components/schemas/RunPreviewSummarySample'
          type: array
          title: Likely Duplicate Pages
        drift_checks:
          items:
            $ref: '#/components/schemas/RunPreviewSummaryDriftCheck'
          type: array
          title: Drift Checks
      type: object
      required:
      - run_id
      - captured_url_count
      - artifacts_count
      - captured_resources_count
      - pdf_count
      - likely_decision_page_count
      - likely_boilerplate_page_count
      - likely_duplicate_page_count
      - content_type_breakdown
      - likely_decision_pages
      - likely_boilerplate_pages
      - likely_duplicate_pages
      - drift_checks
      title: RunPreviewSummaryResponse
...
    RunReadinessResponse:
      properties:
        source_id:
          type: string
          title: Source Id
        source_version_id:
          type: string
          title: Source Version Id
        mode:
          $ref: '#/components/schemas/RunMode'
        ready:
          type: boolean
          title: Ready
        checks:
          items:
            $ref: '#/components/schemas/RunReadinessCheck'
          type: array
          title: Checks
      additionalProperties: false
      type: object
      required:
      - source_id
      - source_version_id
      - mode
      - ready
      - checks
      title: RunReadinessResponse
...
    RunResponse:
      properties:
        run_id:
          type: string
          title: Run Id
          examples:
          - run_01hzxk9c4dnpf8h6t2m5q7w3
        source_id:
          type: string
          title: Source Id
        source_version_id:
          type: string
          title: Source Version Id
        mode:
          $ref: '#/components/schemas/RunMode'
        scope:
          $ref: '#/components/schemas/RunScopeRequest'
        replay:
          anyOf:
          - $ref: '#/components/schemas/RunReplayRequest'
          - type: 'null'
        replay_checkpoint:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Replay Checkpoint
        status:
          $ref: '#/components/schemas/RunStatus'
        started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Started At
        completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Completed At
        artifacts_count:
          type: integer
          title: Artifacts Count
        captured_resources_count:
          type: integer
          title: Captured Resources Count
        failure_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Failure Reason
        refused:
          type: boolean
          title: Refused
          description: True when the ADR-0030 two-key lock blocked this dispatch. The run is a refusal
            record — terminal FAILED, never dispatched — kept so an operator can ask what was attempted
            and why it was refused (#634).
          default: false
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - run_id
      - source_id
      - source_version_id
      - mode
      - scope
      - replay
      - status
      - started_at
      - completed_at
      - artifacts_count
      - captured_resources_count
      - failure_reason
      - created_at
      - updated_at
      title: RunResponse
...
    RunStatus:
      type: string
      enum:
      - pending
      - running
      - completed
      - failed
      - cancelled
      title: RunStatus
...
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
...